Flytter SlettBrev inn i egen fil
